% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Basic.integrand}
\alias{Basic.integrand}
\title{Helper function for computing the posterior mean, posterior variance}
\usage{
Basic.integrand(u, y, k, tau, data.var)
}
\description{
Helper function for computing the posterior mean, posterior variance
}
\keyword{internal}
