% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Basic.y.vec}
\alias{Basic.y.vec}
\title{Helper function for computing the posterior mean, posterior variance}
\usage{
Basic.y.vec(y, tau, k, data.var)
}
\description{
Helper function for computing the posterior mean, posterior variance
}
\keyword{internal}
