% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{coef.horseshoe}
\alias{coef.horseshoe}
\title{Extract Coefficients from a Horseshoe Object}
\usage{
\method{coef}{horseshoe}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"horseshoe"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Named numeric vector of posterior mean coefficients.
}
\description{
Extract Coefficients from a Horseshoe Object
}
