% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsreg.R
\name{coef.hsreg}
\alias{coef.hsreg}
\title{Extract Coefficients from a Horseshoe Object}
\usage{
\method{coef}{hsreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"hsreg"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Named numeric vector of posterior mean coefficients.
}
\description{
Extract Coefficients from a Horseshoe Object
}
