% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsreg-package.R
\docType{package}
\name{hsreg-package}
\alias{hsreg-package}
\title{hsreg: Bayesian Linear Regression with the Horseshoe Prior}
\description{
Gibbs sampler for Bayesian linear regression with the horseshoe prior (Carvalho, Polson & Scott, 2010), supporting selective per-parameter penalization control. Implements the auxiliary-variable sampler from Makalic & Schmidt (2015, \doi{10.48550/arXiv.1508.03884}) with extensions for unpenalized coefficient subsets (flat priors on user-specified parameters), convergence diagnostics via effective sample size (ESS), posterior credible intervals, CATE extraction for multi-arm treatment designs, and full S3 method support (print, summary, predict, coef, vcov, confint). Depends only on base R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amceachin-code/horseshoe}
  \item Report bugs at \url{https://github.com/amceachin-code/horseshoe/issues}
}

}
\author{
\strong{Maintainer}: Andrew McEachin \email{andrew.mceachin@gmail.com}

}
\keyword{internal}
