% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{print.summary.horseshoe}
\alias{print.summary.horseshoe}
\title{Print Method for Summary of Horseshoe Objects}
\usage{
\method{print}{summary.horseshoe}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.horseshoe"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for Summary of Horseshoe Objects
}
