% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsreg.R
\name{print.summary.hsreg}
\alias{print.summary.hsreg}
\title{Print Method for Summary of Horseshoe Objects}
\usage{
\method{print}{summary.hsreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.hsreg"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for Summary of Horseshoe Objects
}
